fixed alerts view related to corrections

+++ app/models/activity.rb
@@ -1,11 +1,13 @@
 
 class Activity < ActiveRecord::Base
-	attr_accessible :action, :trackable, :recipient_id, :user_id, :read
+	attr_accessible :action, :trackable, :recipient_id, :user_id, :read, :position
   	belongs_to :trackable, polymorphic: true
   	belongs_to :user 
 
   	def self.destroy_old_data
-  		delete_all(['created_at < ?', 30.days.ago])
+  		unless self.trackable_type=="Post" && self.action.in?(["qualify", "notify", "at_grade", "below_grade", "publish"])
+  			delete_all(['created_at < ?', 30.days.ago])
+  		end
   	end
